Add files via upload
https://www.coursera.org/specializations/data-analytics-in-the-public-sector-with-r?
https://vita.had.co.nz/papers/layered-grammar.pdf
https://ggplot2-book.org/

In this educational series focused on Advanced Data Science with R, we explore the intricacies of data handling, transformation, and visualization in R, using a dataset provided by the University of Michigan. This series is not just about learning and applying new concepts in R; it's also about streamlining workflows, sharing knowledge, and enhancing understanding of data analytics in the public sector.

This section specifically delves into the manipulation and understanding of categorical data in R, utilizing a dataset on Seattle Pet Licenses. We'll be exploring how to convert data types to categorical, or as they are known in R, factors. This is crucial for effective data visualization and analysis, especially when working with tools like ggplot2 and the tidyverse package.