Add files via upload
The dataset used in this project is not included in the GitHub repository due to its large size. It can be downloaded from the UCI Machine Learning Repository.
The plot*.R scripts are configured to work with the dataset assuming it is located in the same directory as the scripts.